+++ packages/sdk/src/endpoints/AttributesEndpoint.ts
@@ -2,19 +2,34 @@ import {
     ConnectorAttribute,
     ConnectorAttributes,
     ConnectorResponse,
-    CreateAttributeRequest,
+    CreateRepositoryAttributeRequest,
     ExecuteIQLQueryRequest,
     ExecuteIdentityAttributeQueryRequest,
     ExecuteRelationshipAttributeQueryRequest,
     ExecuteThirdPartyRelationshipAttributeQueryRequest,
     GetAttributesRequest,
-    GetValidAttributesRequest
+    GetValidAttributesRequest,
+    NotifyPeerAboutIdentityAttributeSuccessionRequest,
+    NotifyPeerAboutIdentityAttributeSuccessionResponse,
+    SucceedAttributeRequest,
+    SucceedAttributeResponse
 } from "../types";
 import { Endpoint } from "./Endpoint";
 
 export class AttributesEndpoint extends Endpoint {
-    public async createAttribute(request: CreateAttributeRequest): Promise<ConnectorResponse<ConnectorAttribute>> {
+    public async createRepositoryAttribute(request: CreateRepositoryAttributeRequest): Promise<ConnectorResponse<ConnectorAttribute>> {
         return await this.post("/api/v2/Attributes", request);
+    }
+
+    public async succeedAttribute(predecessorId: string, request: SucceedAttributeRequest): Promise<ConnectorResponse<SucceedAttributeResponse>> {
+        return await this.post(`/api/v2/Attributes/${predecessorId}/Succeed`, request);
+    }
+
+    public async notifyPeerAboutIdentityAttributeSuccession(
+        attributeId: string,
+        request: NotifyPeerAboutIdentityAttributeSuccessionRequest
+    ): Promise<ConnectorResponse<NotifyPeerAboutIdentityAttributeSuccessionResponse>> {
+        return await this.post(`/api/v2/Attributes/${attributeId}/NotifyPeer`, request);
     }
 
     public async getAttributes(request: GetAttributesRequest): Promise<ConnectorResponse<ConnectorAttributes>> {