+++ packages/sdk/src/types/attributes/requests/CreateAttributeRequest.ts
@@ -1,5 +1,0 @@
-import { ConnectorIdentityAttribute, ConnectorRelationshipAttribute } from "../ConnectorAttribute";
-
-export interface CreateAttributeRequest {
-    content: ConnectorIdentityAttribute | ConnectorRelationshipAttribute;
-}