+++ packages/sdk/src/types/attributes/requests/CreateRepositoryAttributeRequest.ts
@@ -1,0 +1,10 @@
+import { ConnectorAttributeValue } from "../ConnectorAttribute";
+
+export interface CreateRepositoryAttributeRequest {
+    content: {
+        value: ConnectorAttributeValue;
+        tags?: string[];
+        validFrom?: string;
+        validTo?: string;
+    };
+}