+++ packages/sdk/src/types/attributes/requests/NotifyPeerAboutIdentityAttributeSuccessionRequest.ts
@@ -1,0 +1,3 @@
+export interface NotifyPeerAboutIdentityAttributeSuccessionRequest {
+    peer: string;
+}