+++ packages/sdk/src/types/attributes/requests/NotifyPeerAboutIdentityAttributeSuccessionResponse.ts
@@ -1,0 +1,7 @@
+import { ConnectorAttribute } from "../ConnectorAttribute";
+
+export interface NotifyPeerAboutIdentityAttributeSuccessionResponse {
+    predecessor: ConnectorAttribute;
+    successor: ConnectorAttribute;
+    notificationId: string;
+}