+++ packages/sdk/src/types/attributes/requests/SucceedAttributeRequest.ts
@@ -1,0 +1,10 @@
+import { ConnectorAttributeValue } from "../ConnectorAttribute";
+
+export interface SucceedAttributeRequest {
+    successorContent: {
+        value: ConnectorAttributeValue;
+        tags?: string[];
+        validFrom?: string;
+        validTo?: string;
+    };
+}