+++ test/attributes.test.ts
@@ -1,27 +1,34 @@
-import { ConnectorClient } from "@nmshd/connector-sdk";
+import { ConnectorClient, ConnectorRelationshipAttribute } from "@nmshd/connector-sdk";
 import { Launcher } from "./lib/Launcher";
 import { QueryParamConditions } from "./lib/QueryParamConditions";
-import { createAttribute } from "./lib/testUtils";
+import {
+    createRepositoryAttribute,
+    establishRelationship,
+    executeFullCreateAndShareIdentityAttributeFlow,
+    executeFullCreateAndShareRelationshipAttributeFlow,
+    syncUntilHasMessageWithNotification,
+    syncUntilHasMessages
+} from "./lib/testUtils";
 import { ValidationSchema } from "./lib/validation";
 
 const launcher = new Launcher();
 let client1: ConnectorClient;
+let client2: ConnectorClient;
 let client1Address: string;
+let client2Address: string;
 
 beforeAll(async () => {
-    [client1] = await launcher.launch(1);
+    [client1, client2] = await launcher.launch(2);
+    await establishRelationship(client1, client2);
     client1Address = (await client1.account.getIdentityInfo()).result.address;
+    client2Address = (await client2.account.getIdentityInfo()).result.address;
 }, 30000);
 afterAll(() => launcher.stop());
 
 describe("Attributes", () => {
-    let attributeId: string;
-
-    test("should create an attribute", async () => {
-        const createAttributeResponse = await client1.attributes.createAttribute({
+    test("should create a repository attribute", async () => {
+        const createAttributeResponse = await client1.attributes.createRepositoryAttribute({
             content: {
-                "@type": "IdentityAttribute",
-                owner: client1Address,
                 value: {
                     "@type": "GivenName",
                     value: "AGivenName"
@@ -31,11 +38,20 @@ describe("Attributes", () => {
         });
 
         expect(createAttributeResponse).toBeSuccessful(ValidationSchema.ConnectorAttribute);
-
-        attributeId = createAttributeResponse.result.id;
     });
 
     test("should get the created attribute", async () => {
+        const attributeId = (
+            await client1.attributes.createRepositoryAttribute({
+                content: {
+                    value: {
+                        "@type": "GivenName",
+                        value: "AGivenName"
+                    },
+                    tags: ["content:edu.de"]
+                }
+            })
+        ).result.id;
         const getAttributeResponse = await client1.attributes.getAttribute(attributeId);
         expect(getAttributeResponse).toBeSuccessful(ValidationSchema.ConnectorAttribute);
     });
@@ -49,16 +65,112 @@ describe("Attributes", () => {
         const getAttributesResponse = await client1.attributes.getValidAttributes({});
         expect(getAttributesResponse).toBeSuccessful(ValidationSchema.ConnectorAttributes);
     });
+
+    test("should succeed a Repository Attribute", async () => {
+        const newRepositoryAttribute = {
+            content: {
+                value: {
+                    "@type": "GivenName",
+                    value: "AGivenName"
+                },
+                tags: ["content:edu.de"]
+            }
+        };
+        const createAttributeResponse = await client1.attributes.createRepositoryAttribute(newRepositoryAttribute);
+
+        const attributeId = createAttributeResponse.result.id;
+
+        const succeedAttributeResponse = await client1.attributes.succeedAttribute(attributeId, {
+            successorContent: {
+                value: {
+                    "@type": "GivenName",
+                    value: "AGivenName"
+                },
+                tags: ["content:edu.de"]
+            }
+        });
+
+        expect(succeedAttributeResponse.isSuccess).toBe(true);
+
+        const succeededAttribute = (await client1.attributes.getAttribute(succeedAttributeResponse.result.successor.id)).result;
+
+        expect(succeededAttribute.content).toStrictEqualExcluding(newRepositoryAttribute.content, "@type", "owner");
+    });
+
+    test("Should notify peer about Repository Attribute Succession", async () => {
+        const ownSharedRepositoryAttribute = await executeFullCreateAndShareIdentityAttributeFlow(client1, client2, {
+            "@type": "IdentityAttribute",
+            owner: client1Address,
+            value: {
+                "@type": "GivenName",
+                value: "AGivenName"
+            }
+        });
+
+        expect(ownSharedRepositoryAttribute.shareInfo?.sourceAttribute).toBeDefined();
+
+        const ownUnsharedRepositoryAttribute = await client1.attributes.getAttribute(ownSharedRepositoryAttribute.shareInfo!.sourceAttribute!);
+
+        const successionResponse = await client1.attributes.succeedAttribute(ownUnsharedRepositoryAttribute.result.id, {
+            successorContent: {
+                value: {
+                    "@type": "GivenName",
+                    value: "ANewGivenName"
+                }
+            }
+        });
+        expect(successionResponse.isSuccess).toBe(true);
+        const notificationResponse = await client1.attributes.notifyPeerAboutIdentityAttributeSuccession(successionResponse.result.successor.id, {
+            peer: client2Address
+        });
+        expect(notificationResponse.isSuccess).toBe(true);
+
+        await syncUntilHasMessageWithNotification(client2, notificationResponse.result.notificationId);
+
+        const succeededAttributeResponse = await client2.attributes.getAttribute(notificationResponse.result.successor.id);
+
+        expect(notificationResponse.result.successor.content.value).toStrictEqual(succeededAttributeResponse.result.content.value);
+    });
+
+    test("Should succeed a Relationship Attribute", async () => {
+        const attributeContent: ConnectorRelationshipAttribute = {
+            owner: client1Address,
+            "@type": "RelationshipAttribute",
+            value: {
+                "@type": "ProprietaryString",
+                title: "text",
+                value: "AGivenName"
+            },
+            key: "key",
+            confidentiality: "public"
+        };
+        const attribute = await executeFullCreateAndShareRelationshipAttributeFlow(client1, client2, attributeContent);
+
+        const successionAttribute = {
+            successorContent: {
+                value: {
+                    "@type": "ProprietaryString",
+                    title: "text",
+                    value: "ANewGivenName"
+                }
+            }
+        };
+        const successionResponse = await client1.attributes.succeedAttribute(attribute.id, successionAttribute);
+        expect(successionResponse.isSuccess).toBe(true);
+
+        await syncUntilHasMessages(client2);
+
+        const client2SuccessorResponse = await client2.attributes.getAttribute(successionResponse.result.successor.id);
+
+        expect(client2SuccessorResponse.result.content.value).toStrictEqual(successionAttribute.successorContent.value);
+    });
 });
 
 describe("Attributes Query", () => {
     test("should query attributes", async () => {
-        const client1Address = (await client1.account.getIdentityInfo()).result.address;
         const attribute = (
-            await client1.attributes.createAttribute({
+            await client1.attributes.createRepositoryAttribute({
                 content: {
-                    "@type": "IdentityAttribute",
-                    owner: client1Address,
                     value: {
                         "@type": "GivenName",
                         value: "AGivenName"
@@ -88,12 +200,9 @@ describe("Attributes Query", () => {
     });
 
     test("should query valid attributes", async () => {
-        const client1Address = (await client1.account.getIdentityInfo()).result.address;
         const attribute = (
-            await client1.attributes.createAttribute({
+            await client1.attributes.createRepositoryAttribute({
                 content: {
-                    "@type": "IdentityAttribute",
-                    owner: client1Address,
                     value: {
                         "@type": "GivenName",
                         value: "AGivenName"
@@ -122,10 +231,8 @@ describe("Attributes Query", () => {
 
 describe("Execute AttributeQueries", () => {
     test("should execute an IdentityAttributeQuery", async () => {
-        const attribute = await createAttribute(client1, {
+        const attribute = await createRepositoryAttribute(client1, {
             content: {
-                "@type": "IdentityAttribute",
-                owner: client1Address,
                 value: {
                     "@type": "GivenName",
                     value: "AGivenName"
@@ -133,41 +240,41 @@ describe("Execute AttributeQueries", () => {
             }
         });
 
-        const executeIdentityAttributeQueryResult = await client1.attributes.executeIdentityAttributeQuery({ query: { valueType: "GivenName" } });
-        expect(executeIdentityAttributeQueryResult).toBeSuccessful(ValidationSchema.ConnectorAttributes);
-        const attributes = executeIdentityAttributeQueryResult.result;
+        const executeIdentityAttributeQueryResponse = await client1.attributes.executeIdentityAttributeQuery({ query: { valueType: "GivenName" } });
+        expect(executeIdentityAttributeQueryResponse).toBeSuccessful(ValidationSchema.ConnectorAttributes);
+        const attributes = executeIdentityAttributeQueryResponse.result;
 
         expect(attributes).toContainEqual(attribute);
     });
 
     test("should execute a RelationshipAttributeQuery", async () => {
-        await createAttribute(client1, {
-            content: {
-                "@type": "RelationshipAttribute",
-                owner: client1Address,
-                value: {
-                    "@type": "ProprietaryString",
-                    title: "ATitle",
-                    value: "AString"
-                },
-                key: "AKey",
-                confidentiality: "public"
-            }
-        });
+        const attributeContent: ConnectorRelationshipAttribute = {
+            owner: client1Address,
+            "@type": "RelationshipAttribute",
+            value: {
+                "@type": "ProprietaryString",
+                title: "text",
+                value: "AGivenName"
+            },
+            key: "someSpecialKey",
+            confidentiality: "public"
+        };
+        await executeFullCreateAndShareRelationshipAttributeFlow(client1, client2, attributeContent);
 
-        const executeRelationshipAttributeQueryResult = await client1.attributes.executeRelationshipAttributeQuery({
+        const executeRelationshipAttributeQueryResponse = await client2.attributes.executeRelationshipAttributeQuery({
             query: {
-                key: "AKey",
+                key: "someSpecialKey",
                 owner: client1Address,
                 attributeCreationHints: {
                     valueType: "ProprietaryString",
-                    title: "A title",
+                    title: "text",
                     confidentiality: "public"
                 }
             }
         });
-        expect(executeRelationshipAttributeQueryResult).toBeSuccessful(ValidationSchema.ConnectorAttribute);
 
-        expect(executeRelationshipAttributeQueryResult.result.content.value.value).toBe("AString");
+        expect(executeRelationshipAttributeQueryResponse).toBeSuccessful(ValidationSchema.ConnectorAttribute);
+
+        expect(executeRelationshipAttributeQueryResponse.result.content.value.value).toBe("AGivenName");
     });
 });