+++ test/iqlquery.test.ts
@@ -42,7 +42,7 @@ beforeAll(async () => {
     ];
     attributeIds = [];
     for (const attribute of attributes) {
-        const attributeId = (await client1.attributes.createAttribute({ content: attribute })).result.id;
+        const attributeId = (await client1.attributes.createRepositoryAttribute({ content: { value: attribute.value, tags: attribute.tags } })).result.id;
         attributeIds.push(attributeId);
     }
 
@@ -73,11 +73,7 @@ test("Local IQL Query", async () => {
     ];
 
     for (const e of table) {
-        const queryRequest: ExecuteIQLQueryRequest = {
-            query: {
-                queryString: e.iqlQuery
-            }
-        };
+        const queryRequest: ExecuteIQLQueryRequest = { query: { queryString: e.iqlQuery } };
 
         const response = await client1.attributes.executeIQLQuery(queryRequest);
         const matchedAttributeIds: string[] = response.result.map((e: any) => e.id);
@@ -119,13 +115,7 @@ test("Remote ReadAttributeRequest containing IQL Query", async () => {
     /* Extract and execute IQL query on C1. */
     const incomingRequest = (await client1.incomingRequests.getRequest(requestId)).result;
     const iqlQueryString = ((incomingRequest.content.items[0] as ReadAttributeRequestItem).query as IQLQuery).queryString;
-    const matchedAttributes = (
-        await client1.attributes.executeIQLQuery({
-            query: {
-                queryString: iqlQueryString
-            }
-        })
-    ).result;
+    const matchedAttributes = (await client1.attributes.executeIQLQuery({ query: { queryString: iqlQueryString } })).result;
 
     /* Reply to the response with the first matched attribute. Wait on C2 for
      * the message to arrive. */
@@ -186,13 +176,7 @@ test("Remote ProposeAttributeRequest containing IQL Query with existing attribut
     /* Extract and execute IQL query on C1. */
     const incomingRequest = (await client1.incomingRequests.getRequest(requestId)).result;
     const iqlQueryString = ((incomingRequest.content.items[0] as ReadAttributeRequestItem).query as IQLQuery).queryString;
-    const matchedAttributes = (
-        await client1.attributes.executeIQLQuery({
-            query: {
-                queryString: iqlQueryString
-            }
-        })
-    ).result;
+    const matchedAttributes = (await client1.attributes.executeIQLQuery({ query: { queryString: iqlQueryString } })).result;
 
     /* Reply to the response with the first matched attribute. Wait on C2 for
      * the message to arrive. */
@@ -247,19 +231,13 @@ test("Remote ProposeAttributeRequest containing IQL Query without existing attri
     const incomingRequest = (await client1.incomingRequests.getRequest(requestId)).result;
     const incomingRequestItem: ProposeAttributeRequestItem = incomingRequest.content.items[0] as ProposeAttributeRequestItem;
     const iqlQueryString = (incomingRequestItem.query as IQLQuery).queryString;
-    const matchedAttributes = (
-        await client1.attributes.executeIQLQuery({
-            query: {
-                queryString: iqlQueryString
-            }
-        })
-    ).result;
+    const matchedAttributes = (await client1.attributes.executeIQLQuery({ query: { queryString: iqlQueryString } })).result;
 
     expect(matchedAttributes).toHaveLength(0);
 
     incomingRequestItem.attribute.owner = client1Address;
-
-    const attributeId = (await client1.attributes.createAttribute({ content: incomingRequestItem.attribute })).result.id;
+    const requestItemAttribute = incomingRequestItem.attribute;
+    const attributeId = (await client1.attributes.createRepositoryAttribute({ content: { value: requestItemAttribute.value } })).result.id;
 
     /* Reply to the response with the first matched attribute. Wait on C2 for
      * the message to arrive. */